src/siglip_service/requirements.txt
torch --index-url https://download.pytorch.org/whl/cpu
transformers
pillow
fastapi
uvicorn
python-multipart
einops
sentencepiece
protobuf
